refactor: split by responsibilities
