+++ src/Timeline/PlayControls.tsx
@@ -1,6 +1,7 @@
-import React, { memo, useCallback, useContext, useRef } from "react";
-import { TimelineContext } from "./TimelineContext";
+import { memo, useContext } from "react";
+import { CurrentTimeInput } from "./CurrentTimeInput";
 import { RenderTracker } from "./RenderTracker";
+import { TimelineContext } from "./TimelineContext";
 
 const config = {
   /**
@@ -11,141 +12,9 @@ const config = {
   maxEndTime: 2000,
 };
 
-export function validateTime(rawTime: number): number {
-  let time = Math.round(rawTime / config.timeStep) * config.timeStep;
-  if (Number.isNaN(rawTime) || rawTime === -Infinity) {
-    time = config.minEndTime;
-  } else if (time === Infinity) {
-    time = config.maxEndTime;
-  }
-  // TODO round decimal values
-  return Math.min(config.maxEndTime, Math.max(config.minEndTime, time));
-}
-
-/**
-    How this component work:
-
-    Entry (focus)
-    1. click on the input area
-    2. click on the spinners
-    3. tab
-
-    Change local time only
-    1. type in the input
-
-    Change global time directly
-    1. press enter
-    2. click outside (blur)
-    3. click on the spinners
-
-    Revert to global time
-    1. press escape
- */
 export const PlayControls = () => {
-  const inputElement = useRef<HTMLInputElement>(null);
-  const [isEditingTime, setIsEditingTime] = React.useState(false);
   const { time: globalTime, setTime: setGlobalTime } =
     useContext(TimelineContext);
-  const globalTimeString = globalTime.toString(); // Remove leading zeroes
-  const [inputTime, setInputTime] = React.useState<string>(globalTimeString);
-  const isSpinnerClicked = useRef(false);
-  const isArrowKeyDown = useRef(false);
-  const blurTriggerKey = useRef<string | null>(null);
-
-  const selectInputText = useCallback(() => {
-    inputElement.current?.select();
-  }, []);
-
-  const setGlobalTimeAndSyncInputTime = useCallback((inputTime: string) => {
-    const rawValue = parseFloat(inputTime);
-    const validatedValue = validateTime(rawValue);
-    setGlobalTime(validatedValue);
-    setInputTime(validatedValue.toString());
-  }, []);
-
-  const onInputChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      if (isSpinnerClicked.current || isArrowKeyDown.current) {
-        setGlobalTimeAndSyncInputTime(e.target.value);
-        selectInputText();
-      } else {
-        setInputTime(e.target.value);
-        // TODO make invalid text red
-      }
-    },
-    [selectInputText, setGlobalTimeAndSyncInputTime]
-  );
-
-  const onInputFocus = useCallback(() => {
-    setIsEditingTime(true);
-    selectInputText();
-  }, [selectInputText]);
-
-  const onInputBlur = () => {
-    if (blurTriggerKey.current === "Escape") {
-      setInputTime(globalTimeString);
-    } else if (
-      blurTriggerKey.current === "Enter" ||
-      blurTriggerKey.current === null
-    ) {
-      setGlobalTimeAndSyncInputTime(inputTime);
-    }
-
-    setIsEditingTime(false);
-    blurTriggerKey.current = null;
-  };
-
-  const onInputKeyDown = useCallback(
-    (e: React.KeyboardEvent<HTMLInputElement>) => {
-      /**
-       * If any key is pressed, it would break the spinner event chain, so we clear the flag.
-       */
-      isSpinnerClicked.current = false;
-
-      if (e.key === "Enter" || e.key === "Escape") {
-        blurTriggerKey.current = e.key;
-        e.currentTarget.blur();
-      } else if (e.key === "ArrowUp" || e.key === "ArrowDown") {
-        isArrowKeyDown.current = true;
-      }
-    },
-    []
-  );
-
-  const onInputKeyUp = useCallback(
-    (_e: React.KeyboardEvent<HTMLInputElement>) => {
-      isArrowKeyDown.current = false;
-    },
-    []
-  );
-
-  const onInputMouseDown = useCallback(
-    (_e: React.MouseEvent<HTMLInputElement>) => {
-      /**
-       * The goal is to change the global time when the spinners are clicked.
-       *
-       * However, we cannot know if they are clicked due to the limitation of browser implementation.
-       * The browser can only tell us that the input element was clicked, but it could be any pixel.
-       *
-       * The workaround is to detect a chain of consecutive events: [mouse down, change].
-       * When a spinner was clicked, it would form this event chain.
-       *
-       * So that in this handler, we firstly set the flag to true,
-       * and then we check if the chain is formed in the other handlers.
-       */
-      isSpinnerClicked.current = true;
-    },
-    []
-  );
-
-  const onInputMouseUp = useCallback(
-    (_e: React.MouseEvent<HTMLInputElement>) => {
-      isSpinnerClicked.current = false;
-    },
-    []
-  );
-
-  const timeInputValue = isEditingTime ? inputTime : globalTimeString;
 
   return (
     <>
@@ -157,22 +26,14 @@ export const PlayControls = () => {
       >
         <fieldset className="flex gap-1">
           Current
-          <input
-            className="bg-gray-700 px-1 rounded"
-            type="number"
+          <CurrentTimeInput
+            value={globalTime}
+            onChange={setGlobalTime}
             data-testid="current-time-input"
-            ref={inputElement}
             min={0}
             max={config.maxEndTime}
             step={config.timeStep}
-            value={timeInputValue}
-            onChange={onInputChange}
-            onFocus={onInputFocus}
-            onBlur={onInputBlur}
-            onKeyDown={onInputKeyDown}
-            onKeyUp={onInputKeyUp}
-            onMouseDown={onInputMouseDown}
-            onMouseUp={onInputMouseUp}
+            config={config}
           />
         </fieldset>
         -
